refactor: Header 리팩토링
